requirements.txt
asgiref==3.9.1
Django==5.2.4
django-debug-toolbar==6.0.0
Faker==37.5.3
psycopg-binary==3.2.9
psycopg2==2.9.10
sqlparse==0.5.3
tzdata==2025.2
